Facing Loop problem
i am facing problem while applying a loop in this program.
As the Dw can be more, less or equal to the value available in 2nd column of the spt table in ini.mat file. So i applied "if" condition and two different  formulas for the same.
But it only goes in Dw>=spt(i,2) condition.
i want it to compare the value from every single row of spt table(column 2) with Dw value and apply the required condition.